# apps/api/src/main.ts
/**
 * Headless API Server (REST + GraphQL)
 *
 * This is a multi-protocol API server that exposes the same services via:
 * - REST API (standard HTTP/JSON)
 * - GraphQL API (queries, mutations, subscriptions)
 *
 * Perfect for:
 * - Mobile app backends
 * - External integrations
 * - Multi-language clients
 * - Webhooks and third-party services
 *
 * DRY Architecture:
 * All API styles (REST, GraphQL, tRPC) share the SAME service layer.
 * No code duplication - services are resolved from the DI container.
 *
 * 12-Factor Compliance:
 * ✅ Factor III: Config from environment variables
 * ✅ Factor VII: Port binding - self-contained HTTP server
 * ✅ Factor IX: Graceful shutdown
 * ✅ Factor XI: Logs to stdout/stderr
 */

import {
  bootstrapMicroservice,
  type MicroserviceContext,
  validateEnvironment,
  getEnvironment,
  setupGracefulShutdown,
} from "@network-monitor/infrastructure";
import type {
  IMonitorService,
  IAlertingService,
  INotificationService,
} from "@network-monitor/shared";
import { Hono } from "hono";
import { logger as honoLogger } from "hono/logger";
import { cors } from "hono/cors";
import { createYoga } from "graphql-yoga";
import { makeExecutableSchema } from "@graphql-tools/schema";
import { schema } from "./graphql/schema";
import { createResolvers, createGraphQLContext } from "./graphql/resolvers";
import { registerRESTRoutes } from "./rest/routes";
import {
  getOpenAPISpec,
  handleDocsRequest,
  autoLaunchBrowser,
} from "./openapi/swagger-ui";

async function startMonolith() {
  // 12-Factor Factor III: Validate environment configuration at startup
  validateEnvironment();
  const config = getEnvironment();

  // Justification: Console statements for 12-factor compliance banner
  // eslint-disable-next-line no-console
  console.log("🚀 Starting Network Monitor Monolith (12-Factor)...");
  // eslint-disable-next-line no-console
  console.log("📦 All services in one process");
  // eslint-disable-next-line no-console
  console.log(`🌍 Environment: ${config.nodeEnv}`);
  // eslint-disable-next-line no-console
  console.log(`⚙️  Event Bus: ${config.eventBusType}`);
  // eslint-disable-next-line no-console
  console.log(`📊 Log Level: ${config.logLevel}`);
  // eslint-disable-next-line no-console
  console.log("");

  // Bootstrap with environment configuration
  // Service wiring: Which implementations to use (DI configuration)
  // Runtime config: All values come from environment variables
  const wiringConfig = process.env.SERVICE_WIRING_CONFIG || config.nodeEnv;
  const configPath = `service-wiring/${wiringConfig}.json`;

  // Justification: Console statements for service wiring info
  // eslint-disable-next-line no-console
  console.log(`📦 Service wiring: ${wiringConfig}.json`);
  // eslint-disable-next-line no-console
  console.log(`⚙️  Runtime config: from environment variables`);

  const context = await bootstrapMicroservice({
    serviceName: "Network Monitor Monolith",
    configPath,
    enableDatabase: true,
    showBanner: false, // Custom banner above

    // Custom initialization - setup all services
    onInitialized: async (ctx: MicroserviceContext) => {
      // Get all services
      const monitorService = ctx.services.monitor as IMonitorService | null;
      const alertingService = ctx.services.alerting as IAlertingService | null;
      const notificationService = ctx.services
        .notification as INotificationService | null;

      ctx.logger.info("Infrastructure initialized", {
        eventBus: config.eventBusType,
        database: ctx.database ? "connected" : "not configured",
        services: {
          monitor: config.monitorServiceEnabled,
          alerting: config.alertingServiceEnabled,
          notification: config.notificationServiceEnabled,
        },
      });

      ctx.logger.info("Services running:");
      if (monitorService) {
        ctx.logger.info("  ✅ Monitor Service (targets, speed tests)");
      }
      if (alertingService) {
        ctx.logger.info("  ✅ Alerting Service (alert rules, incidents)");
      }
      if (notificationService) {
        ctx.logger.info("  ✅ Notification Service (push, in-app)");
      }
      ctx.logger.info("");

      ctx.logger.info("12-Factor Compliance:");
      ctx.logger.info("  ✅ Factor III: Config from environment");
      ctx.logger.info("  ✅ Factor IX: Graceful shutdown enabled");
      ctx.logger.info("  ✅ Factor XI: Logs to stdout/stderr");
      ctx.logger.info("  ✅ Factor X: PostgreSQL supported");
      ctx.logger.info("");

      ctx.logger.info("To scale to microservices:");
      ctx.logger.info("  1. Set EVENT_BUS_TYPE=rabbitmq");
      ctx.logger.info("  2. Deploy services separately (apps/*-service/)");
      ctx.logger.info("  3. Zero code changes needed!");
      ctx.logger.info("");

      // 12-Factor Factor IX: Setup graceful shutdown
      setupGracefulShutdown({
        logger: ctx.logger,
        database: ctx.database ?? undefined,
        eventBus: ctx.eventBus,
        shutdownTimeout: 30000,
        onShutdown: async () => {
          ctx.logger.info("Running custom monolith cleanup...");
          // Custom cleanup if needed
        },
      });

      ctx.logger.info(
        "Graceful shutdown handlers registered (SIGTERM, SIGINT)"
      );
    },

    // Custom shutdown - cleanup all services
    onShutdown: async (ctx: MicroserviceContext) => {
      ctx.logger.info("Shutting down all services...");
      // All cleanup is handled by the graceful shutdown handler
    },
  });

  // Load OpenAPI specification
  const openAPISpec = getOpenAPISpec();

  // Create GraphQL server
  const executableSchema = makeExecutableSchema({
    typeDefs: schema,
    resolvers: createResolvers(context),
  });

  const yoga = createYoga({
    schema: executableSchema,
    context: (req: { request: Request }) =>
      createGraphQLContext(req.request, context),
    graphqlEndpoint: "/graphql",
    landingPage: true, // GraphQL Playground
  });

  // Create Hono app
  const app = new Hono();

  // Middleware
  app.use("*", honoLogger()); // Request logging
  app.use(
    "*",
    cors({
      origin: "*", // TODO: Configure allowed origins via environment
      credentials: true,
    })
  );

  // Root - API documentation
  app.get("/", c => {
    return c.json(
      {
        name: "Network Monitor API",
        version: "1.0.0",
        description: "Multi-protocol API server with Hono",
        documentation: {
          openapi: `http://${config.host}:${config.port}/api/docs`,
          graphql: `http://${config.host}:${config.port}/graphql`,
          postman: "Import postman-collection.json",
          insomnia: "Import insomnia-collection.json",
        },
        endpoints: {
          rest: {
            documentation: "/api/docs",
            spec: "/api/docs/spec.yaml",
            baseUrl: "/api",
            health: "/health",
          },
          graphql: {
            endpoint: "/graphql",
            playground: "/graphql (visit in browser)",
          },
        },
        services: {
          monitor: !!context.services.monitor,
          alerting: !!context.services.alerting,
          notification: !!context.services.notification,
        },
      },
      200
    );
  });

  // OpenAPI Documentation (Swagger UI)
  app.get("/api/docs", c => handleDocsRequest(c.req.raw, openAPISpec));
  app.get("/api/docs/*", c => handleDocsRequest(c.req.raw, openAPISpec));

  // GraphQL endpoint
  app.all("/graphql", c => yoga.fetch(c.req.raw));

  // Register REST routes
  registerRESTRoutes(app, context);

  // Error handling middleware
  app.onError((err, c) => {
    context.logger.error("Request error", { error: err, path: c.req.path });
    return c.json(
      {
        error: "Internal Server Error",
        message: err.message,
        path: c.req.path,
      },
      500
    );
  });

  // 404 handler
  app.notFound(c => {
    return c.json(
      {
        error: "Not Found",
        path: c.req.path,
        method: c.req.method,
        availableEndpoints: [
          "GET /",
          "GET /health",
          "GET /api/docs",
          "GET /graphql",
          "GET /api/targets",
          "POST /api/targets",
          "GET /api/targets/:id",
          "PUT /api/targets/:id",
          "DELETE /api/targets/:id",
          "POST /api/targets/:id/start",
          "POST /api/targets/:id/stop",
          "POST /api/targets/:id/test",
          "GET /api/targets/active",
          "GET /api/alert-rules/target/:targetId",
          "POST /api/alert-rules",
          "PUT /api/alert-rules/:id",
          "DELETE /api/alert-rules/:id",
          "GET /api/incidents/target/:targetId",
          "POST /api/incidents/:id/resolve",
          "GET /api/notifications/user/:userId",
          "POST /api/notifications/:id/read",
          "POST /api/push-subscriptions",
        ],
      },
      404
    );
  });

  // Start HTTP server using Bun with Hono
  // Justification: Bun global is provided by Bun runtime, not in standard TypeScript types
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  const BunGlobal = globalThis as any;
  const server = BunGlobal.Bun.serve({
    port: config.port,
    hostname: config.host,
    fetch: app.fetch, // Hono handles all routing
  });

  context.logger.info("🚀 API Server is now running", {
    port: config.port,
    host: config.host,
    environment: config.nodeEnv,
    endpoints: {
      rest: `http://${config.host}:${config.port}/api`,
      graphql: `http://${config.host}:${config.port}/graphql`,
      health: `http://${config.host}:${config.port}/health`,
      docs: `http://${config.host}:${config.port}/api/docs`,
    },
  });

  context.logger.info("📚 Available Protocols:");
  context.logger.info("  ✅ REST API - Standard HTTP/JSON");
  context.logger.info("  ✅ GraphQL - Flexible queries and mutations");
  context.logger.info("");

  context.logger.info("📖 Documentation:");
  context.logger.info(
    `  📘 OpenAPI/Swagger: http://${config.host}:${config.port}/api/docs`
  );
  context.logger.info(
    `  🎮 GraphQL Playground: http://${config.host}:${config.port}/graphql`
  );
  context.logger.info(
    `  📄 Postman Collection: apps/api/postman-collection.json`
  );
  context.logger.info(
    `  📄 Insomnia Collection: apps/api/insomnia-collection.json`
  );
  context.logger.info("");

  // Auto-launch browser in development mode
  if (config.nodeEnv === "development") {
    const docsUrl = `http://${config.host}:${config.port}/api/docs`;
    autoLaunchBrowser(docsUrl, context.logger).catch(() => {
      // Silently ignore errors
    });
  }

  // Return server handle for shutdown
  return { context, server };
}

// Start the monolith
startMonolith().catch(error => {
  // Justification: Console error for startup failure
  // eslint-disable-next-line no-console
  console.error("❌ Failed to start monolith:", error);
  process.exit(1);
});